refactor: todoList를 todoContent로 변경 및 구조 리펙토링 기존에는 TodoList외의 태그 요소들도 존재했다. 따라서, 조금더 포괄적인 의미의 TodoContent로 변경
